Feature/yaml

+++ gh_labels.py
@@ -2,12 +2,17 @@
 """Populate GitHub labels for issue tracker."""
 from collections import namedtuple
 import codecs
+import yaml
 import json
 import sys
 import os
 import re
 import requests
 import urllib.parse
+try:
+    from yaml import CLoader as Loader
+except ImportError:
+    from yaml import Loader
 
 __version__ = "1.0.0"
 
@@ -221,6 +226,11 @@ class GhLabelSync:
                 raise ValueError("The name '{}' is already present in the label list".format(name))
             self.labels[name] = value
 
+        self.ignores = set()
+        for name in config.get('ignores', []):
+            self._validate_str(name)
+            self.ignores.add(name.lower())
+
     def _resolve_color(self, color):
         """Parse color."""
 
@@ -278,7 +288,7 @@ class GhLabelSync:
                 updated.add(edit.old)
                 updated.add(edit.new)
             else:
-                if edit is None and self.delete:
+                if edit is None and self.delete and label['name'].lower() not in self.ignores:
                     print('    Deleting {}: #{} "{}"'.format(label['name'], label['color'], label['description']))
                     if not self.debug:
                         self.git.delete_label(label['name'])
@@ -297,6 +307,14 @@ class GhLabelSync:
 
 def main():
     """Main."""
+
+    dbg = os.getenv("INPUT_DEBUG", 'disable')
+    if dbg == 'enable':
+        debug = True
+    elif dbg == 'disable':
+        debug = False
+    else:
+        raise ValueError('Unknown value for debug: {}'.format(dbg))
 
     # Parse mode
     mode = os.getenv("INPUT_MODE", 'normal')
@@ -321,12 +339,13 @@ def main():
 
     # Parse label file
     labels = os.getenv("INPUT_FILE", '.github/labels.json')
+    print('Reading labels from {}'.format(labels))
     with codecs.open(labels, 'r', encoding='utf-8') as f:
-        config = json.loads(f.read())
+        config = yaml.load(f.read(), Loader=Loader)
 
     # Sync the labels
     git = Api(token, user, repo)
-    GhLabelSync(config, git, delete).sync()
+    GhLabelSync(config, git, delete, debug).sync()
     return 0
